+++ apps/openmw/mwgui/travelwindow.cpp
@@ -258,7 +258,7 @@ namespace MWGui
     {
         if (arg.button == SDL_CONTROLLER_BUTTON_A)
         {
-            if (mControllerFocus >= 0 && mControllerFocus < mDestinationButtons.size())
+            if (mControllerFocus >= 0 && mControllerFocus < static_cast<int>(mDestinationButtons.size()))
             {
                 onTravelButtonClick(mDestinationButtons[mControllerFocus]);
                 MWBase::Environment::get().getWindowManager()->playSound(ESM::RefId::stringRefId("Menu Click"));
